requirements.txt
flask
pymongo
python-magic